# docs/swagger.yaml
basePath: /api
definitions:
  ApiResponse:
    description: Generic API response
    properties:
      data: {}
      responseKey:
        type: string
    type: object
...
  TargetCount:
    description: TargetCount represents the count of target.
    properties:
      count:
        description: Count is the number of target.
        type: integer
    type: object
  TargetDetails:
    description: TargetDetails represents detailed information about a specific target.
    properties:
      createdAt:
        description: CreatedAt is the timestamp indicating when the service entry
          was created.
        type: string
      description:
        description: description is the description of the service. It is required
          and should be between 3 and 35 characters.
        maxLength: 35
        minLength: 3
        type: string
      id:
        description: ID is the unique identifier for the user.
        type: string
      name:
        description: name is the  name of the service. It is required and should be
          between 3 and 30 characters.
        maxLength: 30
        minLength: 3
        type: string
    required:
    - description
    - name
    type: object
  TargetIn:
    description: TargetIn represents the input structure for creating a new target.
    properties:
      companyID:
        description: CompanyID is the unique identifier for the company associated
          with the service. It is required.
        type: string
      name:
        description: name is the  name of the service. It is required and should be
          between 3 and 30 characters.
        maxLength: 30
        minLength: 3
        type: string
    required:
    - companyID
    - name
    type: object
  TargetList:
    description: TargetList represents a simplified version of the target for listing
      purposes.
    properties:
      id:
        description: ID is the unique identifier for the service.
        type: string
      name:
        description: Name is the full name of the service.
        type: string
    type: object
  TargetPagination:
    description: ServicesPagination represents the paginated list of Services.
    properties:
      items:
        description: Items is a slice containing individual service details.
        items:
          $ref: '#/definitions/TargetTable'
        type: array
      limit:
        description: Limit is the maximum number of items per page in the pagination.
        type: integer
      page:
        description: Page is the current page number in the pagination.
        type: integer
      totalCount:
        description: TotalCount is the total number of services in the entire list.
        type: integer
    type: object
  TargetTable:
    description: TargetTable represents a single target entry in a table.
    properties:
      createdAt:
        description: CreatedAt is the timestamp indicating when the service entry
          was created.
        type: string
      description:
        description: description is the description of the service. It is required
          and should be between 3 and 35 characters.
        maxLength: 35
        minLength: 3
        type: string
      id:
        description: ID is the unique identifier for the user.
        type: string
      name:
        description: name is the  name of the service. It is required and should be
          between 3 and 30 characters.
        maxLength: 30
        minLength: 3
        type: string
    required:
    - description
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: h.manai@tunisiancloud.com
    name: Hedi Manai
  description: A basic template for managing your project using Golang REST APIs.
  title: Project Name
  version: "0.1"
paths:
...
  /target/{companyID}:
    get:
      description: Get all targets.
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TargetPagination'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get targets
      tags:
      - Target
    post:
      consumes:
      - application/json
      description: Create a new target.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: target query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/TargetIn'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Create target
      tags:
      - Target
  /target/{companyID}/count:
    get:
      description: Get number of all targets.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TargetCount'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get number of  targets
      tags:
      - Target
  /target/{companyID}/list:
    get:
      description: Get list of all targets.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/TargetList'
            type: array
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get list of  targets
      tags:
      - Target
  /targets/{companyID}/{ID}:
    delete:
      description: Delete one target.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: target ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete target
      tags:
      - Target
    get:
      description: Get one target.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: Target ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TargetDetails'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get target
      tags:
      - Target
    put:
      consumes:
      - application/json
      description: Update target.
      parameters:
      - description: Company ID
        in: path
        name: companyID
        required: true
        type: string
      - description: Target ID
        in: path
        name: ID
        required: true
        type: string
      - description: Target query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/TargetIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Update target
      tags:
      - Target
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"